use custom Chart JS options
