+++ public/mix-manifest.json
@@ -1,6 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=68579da961a68bee0cda",
-    "/css/app.css": "/css/app.css?id=5ce9973b1bc9f6a46cb2",
-    "/js/app.js.map": "/js/app.js.map?id=081aff4a51d6c4aaf839",
-    "/css/app.css.map": "/css/app.css.map?id=5d0439ebaab1434c7ea0"
+    "/js/app.js": "/js/app.js?id=8d5f9d37b9beee7ad7b6",
+    "/css/app.css": "/css/app.css?id=a6e71e015678edc963c5"
 }
